requirements.txt
streamlit
crewai
langchain-groq
python-dotenv
PyPDF2
python-docx